Disable terminal plugin
MATE Terminal now uses GSettings directly via libdconf because it stores
a variable number of profiles.  Accessing this in the terminal plugin
would require using gi.repository.dconf but that cannot be done while
pluma itself uses the static pygobject bindings.  So, this is going to
have to wait until pluma is ported to gobject-introspection.